Move congestion control into its own file
More files = gooder
